// pslpwcl/src/app/modules/pslpapi/model/documeRichiesti.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CheckAutocer } from './checkAutocer';
import { CheckControl } from './checkControl';
import { Cpi } from './cpi';
import { LavAnagrafica } from './lavAnagrafica';
import { MotivoSilp } from './motivoSilp';
import { Operatore } from './operatore';
import { Otp } from './otp';
import { StatoDocume } from './statoDocume';
import { TipoDocume } from './tipoDocume';

export interface DocumeRichiesti { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    dataAccettazione?: Date;
    dataInvio?: Date;
    dataPresaInCarico?: Date;
    dataRifiuto?: Date;
    dataScaduto?: Date;
    idSilTFontedato?: number;
    idSilwebTDocumeRichiesti?: number;
    numRepertorio?: string;
    rifLibra?: string;
    silLavAnagrafica?: LavAnagrafica;
    silOperatore?: Operatore;
    silTCpiEntePromotore?: Cpi;
    silTCpiOperatore?: Cpi;
    silTOtp?: Otp;
    silwebDMotivo?: MotivoSilp;
    silwebTCheckAutocer?: CheckAutocer;
    silwebTCheckControl?: CheckControl;
    silwebTStatoDocume?: StatoDocume;
    silwebTTipoDocume?: TipoDocume;
    timestamp?: number;
}