// pslpwcl/src/app/modules/pslpapi/model/formRicercaRichiesteDocumenti.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CallInfo } from './callInfo';
import { OrderField } from './orderField';

export interface FormRicercaRichiesteDocumenti { 
    callInfo?: CallInfo;
    codiceFiscale?: string;
    cognome?: string;
    dataRichiesteA?: Date;
    dataRichiesteDa?: Date;
    dataStato?: Date;
    idCpiOperatore?: number;
    idEntePromotore?: number;
    idOperatore?: string;
    idSilLavAnagrafica?: number;
    nome?: string;
    ordinamento?: OrderField;
    statoRichiesta?: Array<number>;
    tipoDocumentoRichiesto?: number;
}