// pslpwcl/src/app/modules/pslpapi/model/ricercaRichiesteDocumentiResponse.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiMessage } from './apiMessage';
import { DocumeRichiesti } from './documeRichiesti';

export interface RicercaRichiesteDocumentiResponse { 
    apiMessages?: Array<ApiMessage>;
    currentPage?: number;
    esitoPositivo?: boolean;
    list?: Array<DocumeRichiesti>;
    recordOfPage?: number;
    recordPage?: number;
    totalPage?: number;
    totalResult?: number;
}